L-Variant/consumers/mongodb/requirements.txt
pymongo==3.11.0
kafka-python~=1.4.7
faust